✨ Feat: 회원 목록 조회 기능 구현
회원 멤버십 별로 필터링 기능 추가

+++ src/main/java/com/telegro/telegro/domain/user/entity/User.java
@@ -45,7 +45,7 @@ public class User extends BaseEntity {
     private String zipCode;
 
     @Column
-    private double totalPrice;
+    private Long totalPrice;
 
     @OneToMany(mappedBy = "user", cascade = CascadeType.ALL, orphanRemoval = true)
     private List<DeliveryAddress> deliveryAddresses;
